The architecture of the framework has been redesigned. The user now uses the BluetoothManager to connect/accept bluetooth connections, contrary to implementing his own loaders.
The threads of the connections are held and maintained in the SocketManagerService. Bluetooth messages are sent/received through this service. The Application class has been removed.

BluetoothMessage class has been introduced. Other than the message that the users wants to send and the app code, it also holds information about the sender, the target and if the message is global or private. The class is accompanied by methods to easily create/retrieve them.

Private messages have been added! Just type <\w "device name" your message>, where "device name" is the bluetooth name of the device, and "message" the message that you want to send. I haven't tested what happens if two devices in the chat room have the same name.